more rending app trigger fix
for hukai telebei quest 2
both desc and handin appear at the same time, after the fix the desc now only appear as a placeholder when there is art contest going on or when player lose the art contest. (maybe add a tooltip if the player have to wait for the cooldown)
for hengu bowu quest 2
both desc and handin appear at the same time when player fulfilled the trigger, im not sure if my fix is right.