A new var_effect for adding gov reform power. Using this new function for the battles won effect to avoid too many iterations of the loop. It's okay that you get fractions. Highlight number 3 in an event loc. The disaster events prior to the revolt now check you don't have the revolt flag, just to make sure, Checking for any festival modifier and gaining a random festival modifier have been moved to scripted trigger and scripted efffect respectively. Check you've got 25 mana before allowing you to spend it. Remove the cores of the revolt tags if you beat them. The scripted effect for the revolt is now simpler and is allowed to overshoot the target slightly. To compensate, I reduce the target dev by 4.
Create a new scripted effect for the monthly Ashhana blood curse control. Use a scripted effect to check the thresholds and fire the events, rather than firing a controller event.
Add claw strength from battles before increasing gov power as that reduces the variable to 0.
Create a scripted effect to remove all Zaradyrra festival buffs.
Make sure Adzalan/Shyarayan don't exist for the formable decision, and make sure at least someone has had the revolt (to prevent AI from sniping it before the player).
Update loc for the disaster event to explain that the revolt will only happen if it hasn't happened to anyone else.
Add a note in the loc that the disaster will end if the war isn't over in 5 years.
Tweak Control bar tooltip for casaulties.
New tooltip for the formable decision.
New loc for the disaster ending to handle the case where the revolt has happened to someone else.
Drop the required warscore threshold to 90 in case another nation snipes a province.
Add another end condition if someone annexes the revolt tag while you're at war with them.
Revolt won't happen to the AI if it's at war with a player. Also, the revolt can only happen once so check the global flag.
Trigger .202 5 years after the war starts to trigger the end of the disaster.
Change end event triggers to check 0 warscore since it might happen before you get 90 (just find the winner).
Try to clear most country flags that aren't needed.
If you lose the war, your subjects become sujects of the revolt.
If Adzalan or Shyarayan exist and your disaster ends passively, they no longer annex you.

The disaster uses a different warscore check to handle the case when the revolt tag is vassalised by a third party.
Some comments on the gov mech gfx.
Use 5 instead of 4 as the target dev adjustment when spawning the revolt tag.
Update the loc slightly for the winning event, since it might not be the faction you supported.
The even that spawns the revolt tag sets some flags later so the disaster doesn't end immediately. The backup 5 year event is after, not immediate.
The disaster end event now handles the different cases, when you beat them, they beat you, whether an overlord is involved, and whether they stopped existing.